move getEnabledResolutions, safer replace of imported extension
