+++ server/helpers/utils.ts
@@ -86,16 +86,6 @@ function getUUIDFromFilename (filename: string) {
   return result[0]
 }
 
-function getEnabledResolutions (type: 'vod' | 'live') {
-  const transcoding = type === 'vod'
-    ? CONFIG.TRANSCODING
-    : CONFIG.LIVE.TRANSCODING
-
-  return Object.keys(transcoding.RESOLUTIONS)
-               .filter(key => transcoding.ENABLED && transcoding.RESOLUTIONS[key] === true)
-               .map(r => parseInt(r, 10))
-}
-
 // ---------------------------------------------------------------------------
 
 export {
@@ -105,6 +95,5 @@ export {
   getSecureTorrentName,
   getServerCommit,
   generateVideoImportTmpPath,
-  getUUIDFromFilename,
-  getEnabledResolutions
+  getUUIDFromFilename
 }
